Refactored TreeRTGPinaceaePinus.
